Extract macOS app and window support
