+++ clients/apple/AlanNative/App/AlanMacAppDelegate.swift
@@ -1,0 +1,13 @@
+import AppKit
+
+#if os(macOS)
+final class AlanMacAppDelegate: NSObject, NSApplicationDelegate {
+    func applicationShouldHandleReopen(
+        _ sender: NSApplication,
+        hasVisibleWindows flag: Bool
+    ) -> Bool {
+        AlanMacPrimaryWindowPresenter.focusExistingWindow()
+        return true
+    }
+}
+#endif
