+++ clients/apple/AlanNative/MacShellRootView.swift
@@ -1,203 +1,6 @@
 import SwiftUI
 
 #if os(macOS)
-import AppKit
-
-private extension Color {
-    static func shellAdaptive(
-        light: (Double, Double, Double),
-        dark: (Double, Double, Double),
-        alpha: Double = 1
-    ) -> Color {
-        shellAdaptive(light: light, lightAlpha: alpha, dark: dark, darkAlpha: alpha)
-    }
-
-    static func shellAdaptive(
-        light: (Double, Double, Double),
-        lightAlpha: Double,
-        dark: (Double, Double, Double),
-        darkAlpha: Double
-    ) -> Color {
-        Color(
-            NSColor(name: nil) { appearance in
-                let isDark = appearance.bestMatch(from: [.darkAqua, .aqua]) == .darkAqua
-                let rgb = isDark ? dark : light
-                let alpha = isDark ? darkAlpha : lightAlpha
-                return NSColor(red: rgb.0, green: rgb.1, blue: rgb.2, alpha: alpha)
-            }
-        )
-    }
-}
-
-enum ShellPalette {
-    static let canvas = Color.shellAdaptive(
-        light: (0.94, 0.94, 0.965),
-        dark: (0.045, 0.050, 0.062)
-    )
-    static let window = Color.shellAdaptive(
-        light: (0.972, 0.973, 0.985),
-        dark: (0.055, 0.061, 0.074)
-    )
-    static let sidebar = Color.shellAdaptive(
-        light: (0.922, 0.924, 0.953),
-        dark: (0.071, 0.079, 0.096)
-    )
-    static let sidebarRail = Color.shellAdaptive(
-        light: (0.902, 0.907, 0.941),
-        dark: (0.083, 0.092, 0.112)
-    )
-    static let sidebarCard = Color.shellAdaptive(
-        light: (0.98, 0.98, 0.995),
-        lightAlpha: 1.0,
-        dark: (0.172, 0.188, 0.224),
-        darkAlpha: 0.92
-    )
-    static let sidebarSelection = Color.shellAdaptive(
-        light: (1.0, 1.0, 1.0),
-        lightAlpha: 0.15,
-        dark: (0.205, 0.225, 0.270),
-        darkAlpha: 0.72
-    )
-    static let sidebarHover = Color.shellAdaptive(
-        light: (1.0, 1.0, 1.0),
-        lightAlpha: 0.08,
-        dark: (0.185, 0.205, 0.245),
-        darkAlpha: 0.46
-    )
-    static let sidebarControl = Color.shellAdaptive(
-        light: (1.0, 1.0, 1.0),
-        lightAlpha: 0.12,
-        dark: (0.190, 0.210, 0.252),
-        darkAlpha: 0.54
-    )
-    static let sidebarControlStrong = Color.shellAdaptive(
-        light: (1.0, 1.0, 1.0),
-        lightAlpha: 0.18,
-        dark: (0.215, 0.235, 0.282),
-        darkAlpha: 0.72
-    )
-    static let railBase = Color.shellAdaptive(
-        light: (1.0, 1.0, 1.0),
-        lightAlpha: 0.08,
-        dark: (0.155, 0.172, 0.210),
-        darkAlpha: 0.58
-    )
-    static let railHover = Color.shellAdaptive(
-        light: (1.0, 1.0, 1.0),
-        lightAlpha: 0.14,
-        dark: (0.190, 0.210, 0.252),
-        darkAlpha: 0.66
-    )
-    static let railSelection = Color.shellAdaptive(
-        light: (1.0, 1.0, 1.0),
-        lightAlpha: 0.22,
-        dark: (0.235, 0.255, 0.310),
-        darkAlpha: 0.78
-    )
-    static let workspace = Color.shellAdaptive(
-        light: (0.979, 0.98, 0.989),
-        dark: (0.050, 0.056, 0.070)
-    )
-    static let terminal = Color.shellAdaptive(
-        light: (0.10, 0.12, 0.16),
-        dark: (0.050, 0.061, 0.076)
-    )
-    static let terminalSoft = Color.shellAdaptive(
-        light: (0.16, 0.18, 0.24),
-        dark: (0.100, 0.116, 0.145)
-    )
-    static let accent = Color.shellAdaptive(
-        light: (0.31, 0.39, 0.71),
-        dark: (0.50, 0.60, 0.94)
-    )
-    static let accentSoft = Color.shellAdaptive(
-        light: (0.90, 0.92, 0.98),
-        dark: (0.18, 0.22, 0.34)
-    )
-    static let ink = Color.shellAdaptive(
-        light: (0.16, 0.18, 0.24),
-        dark: (0.90, 0.92, 0.96)
-    )
-    static let mutedInk = Color.shellAdaptive(
-        light: (0.43, 0.45, 0.54),
-        dark: (0.64, 0.68, 0.75)
-    )
-    static let line = Color.shellAdaptive(
-        light: (0.82, 0.83, 0.89),
-        dark: (0.255, 0.285, 0.345)
-    )
-    static let panel = Color.shellAdaptive(
-        light: (1.0, 1.0, 1.0),
-        lightAlpha: 0.74,
-        dark: (0.135, 0.150, 0.180),
-        darkAlpha: 0.78
-    )
-    static let panelSoft = Color.shellAdaptive(
-        light: (1.0, 1.0, 1.0),
-        lightAlpha: 0.60,
-        dark: (0.125, 0.140, 0.170),
-        darkAlpha: 0.64
-    )
-    static let materialScrim = Color.shellAdaptive(
-        light: (0.922, 0.924, 0.953),
-        lightAlpha: 0.35,
-        dark: (0.030, 0.037, 0.050),
-        darkAlpha: 0.78
-    )
-    static let materialTopWash = Color.shellAdaptive(
-        light: (1.0, 1.0, 1.0),
-        lightAlpha: 0.06,
-        dark: (0.130, 0.150, 0.205),
-        darkAlpha: 0.18
-    )
-    static let materialBottomShade = Color.shellAdaptive(
-        light: (0.84, 0.86, 0.92),
-        lightAlpha: 0.04,
-        dark: (0.012, 0.016, 0.024),
-        darkAlpha: 0.34
-    )
-    static let attention = Color.shellAdaptive(
-        light: (0.82, 0.55, 0.24),
-        dark: (0.94, 0.68, 0.34)
-    )
-}
-
-enum ShellRadii {
-    static let control: CGFloat = 6
-    static let row: CGFloat = 8
-    static let surface: CGFloat = 10
-    static let overlay: CGFloat = 12
-}
-
-private struct SidebarMaterialView: NSViewRepresentable {
-    func makeNSView(context: Context) -> NSVisualEffectView {
-        let view = NSVisualEffectView()
-        view.material = .sidebar
-        view.blendingMode = .behindWindow
-        view.state = .followsWindowActiveState
-        return view
-    }
-
-    func updateNSView(_ nsView: NSVisualEffectView, context: Context) {}
-}
-
-struct ShellMaterialBackgroundView: View {
-    var body: some View {
-        ZStack {
-            SidebarMaterialView()
-            ShellPalette.materialScrim
-            LinearGradient(
-                colors: [
-                    ShellPalette.materialTopWash,
-                    ShellPalette.materialBottomShade,
-                ],
-                startPoint: .topLeading,
-                endPoint: .bottomTrailing
-            )
-        }
-    }
-}
-
 struct MacShellRootView: View {
     @ObservedObject private var host: ShellHostController
     @State private var isCommandTabPresented = false
@@ -255,206 +58,6 @@ struct MacShellRootView: View {
         }
         .animation(.easeOut(duration: 0.18), value: isCommandTabPresented)
         .background(ShellWindowPlacementView(metrics: $windowChromeMetrics))
-    }
-}
-
-private struct ShellWindowPlacementView: NSViewRepresentable {
-    @Binding var metrics: ShellWindowChromeMetrics
-
-    func makeNSView(context: Context) -> ShellWindowPlacementNSView {
-        let metricsBinding = _metrics
-        return ShellWindowPlacementNSView { metrics in
-            let newMetrics = metrics
-            DispatchQueue.main.async {
-                guard metricsBinding.wrappedValue != newMetrics else { return }
-                metricsBinding.wrappedValue = newMetrics
-            }
-        }
-    }
-
-    func updateNSView(_ nsView: ShellWindowPlacementNSView, context: Context) {
-        let metricsBinding = _metrics
-        nsView.updateMetricsHandler { metrics in
-            let newMetrics = metrics
-            DispatchQueue.main.async {
-                guard metricsBinding.wrappedValue != newMetrics else { return }
-                metricsBinding.wrappedValue = newMetrics
-            }
-        }
-        nsView.resolveWindowIfNeeded()
-    }
-}
-
-private struct ShellWindowChromeMetrics: Equatable {
-    var trafficLightsTopInset: CGFloat = 0
-}
-
-private final class ShellWindowPlacementNSView: NSView {
-    private var metricsHandler: (ShellWindowChromeMetrics) -> Void
-    private var lastPublishedMetrics: ShellWindowChromeMetrics?
-
-    init(metricsHandler: @escaping (ShellWindowChromeMetrics) -> Void) {
-        self.metricsHandler = metricsHandler
-        super.init(frame: .zero)
-    }
-
-    @available(*, unavailable)
-    required init?(coder: NSCoder) {
-        fatalError("init(coder:) has not been implemented")
-    }
-
-    override func viewDidMoveToWindow() {
-        super.viewDidMoveToWindow()
-        resolveWindowIfNeeded()
-    }
-
-    override func viewDidMoveToSuperview() {
-        super.viewDidMoveToSuperview()
-        resolveWindowIfNeeded()
-    }
-
-    func updateMetricsHandler(_ handler: @escaping (ShellWindowChromeMetrics) -> Void) {
-        metricsHandler = handler
-    }
-
-    func resolveWindowIfNeeded() {
-        DispatchQueue.main.async { [weak self] in
-            guard let window = self?.window else { return }
-            AlanShellWindowPlacement.apply(to: window)
-            let metrics = AlanShellWindowPlacement.chromeMetrics(for: window)
-            guard self?.lastPublishedMetrics != metrics else { return }
-            self?.lastPublishedMetrics = metrics
-            self?.metricsHandler(metrics)
-        }
-    }
-}
-
-private enum AlanShellWindowPlacement {
-    private static var positionedWindowNumbers: Set<Int> = []
-
-    static func apply(to window: NSWindow) {
-        window.title = "Alan"
-        window.isMovableByWindowBackground = true
-        window.minSize = NSSize(width: 1180, height: 760)
-        window.tabbingMode = .disallowed
-
-        if positionedWindowNumbers.insert(window.windowNumber).inserted {
-            window.setFrame(centeredFrameOnMainScreen(for: window), display: true)
-        } else if shouldResetFrame(window.frame) {
-            window.setFrame(centeredFrame(for: window), display: true)
-        }
-
-        if !window.isVisible {
-            window.makeKeyAndOrderFront(nil)
-        }
-
-        NSApp.activate(ignoringOtherApps: true)
-    }
-
-    static func chromeMetrics(for window: NSWindow) -> ShellWindowChromeMetrics {
-        let buttonFrames = [
-            NSWindow.ButtonType.closeButton,
-            .miniaturizeButton,
-            .zoomButton,
-        ].compactMap { buttonType -> NSRect? in
-            window.standardWindowButton(buttonType)?.frame
-        }
-
-        guard let firstFrame = buttonFrames.first else {
-            return ShellWindowChromeMetrics()
-        }
-
-        let trafficLightsFrame = buttonFrames.dropFirst().reduce(firstFrame) { partialResult, frame in
-            partialResult.union(frame)
-        }
-        let topInset = max(0, trafficLightsFrame.maxY + 10)
-
-        return ShellWindowChromeMetrics(
-            trafficLightsTopInset: ceil(topInset)
-        )
-    }
-
-    private static func shouldResetFrame(_ frame: NSRect) -> Bool {
-        let screens = NSScreen.screens
-        guard !screens.isEmpty else { return false }
-
-        let frameArea = max(frame.width * frame.height, 1)
-        let largestVisibleArea = screens
-            .map(\.visibleFrame)
-            .map { visibleFrame -> CGFloat in
-                let intersection = visibleFrame.intersection(frame)
-                guard !intersection.isNull else { return 0 }
-                return intersection.width * intersection.height
-            }
-            .max() ?? 0
-
-        let visibleRatio = largestVisibleArea / frameArea
-        let targetVisibleFrame = preferredVisibleFrame(for: frame)
-
-        return visibleRatio < 0.55
-            || frame.width > targetVisibleFrame.width
-            || frame.height > targetVisibleFrame.height
-            || !targetVisibleFrame.insetBy(dx: -48, dy: -48).intersects(frame)
-    }
-
-    private static func centeredFrame(for window: NSWindow) -> NSRect {
-        let visibleFrame = preferredVisibleFrame(for: window.frame)
-        return centeredFrame(in: visibleFrame, minSize: window.minSize)
-    }
-
-    private static func centeredFrameOnMainScreen(for window: NSWindow) -> NSRect {
-        let visibleFrame = primaryVisibleFrame() ?? preferredVisibleFrame(for: window.frame)
-        return centeredFrame(in: visibleFrame, minSize: window.minSize)
-    }
-
-    private static func centeredFrame(in visibleFrame: NSRect, minSize: NSSize) -> NSRect {
-        let targetWidth = min(
-            max(visibleFrame.width * 0.84, minSize.width),
-            visibleFrame.width - 32
-        )
-        let targetHeight = min(
-            max(visibleFrame.height * 0.86, minSize.height),
-            visibleFrame.height - 32
-        )
-
-        let origin = CGPoint(
-            x: visibleFrame.midX - (targetWidth / 2),
-            y: visibleFrame.midY - (targetHeight / 2)
-        )
-
-        return NSRect(origin: origin, size: NSSize(width: targetWidth, height: targetHeight))
-    }
-
-    private static func preferredVisibleFrame(for frame: NSRect) -> NSRect {
-        if let containingScreen = NSScreen.screens.first(where: {
-            !$0.visibleFrame.intersection(frame).isNull
-        }) {
-            return containingScreen.visibleFrame
-        }
-
-        if let mainFrame = NSScreen.main?.visibleFrame {
-            return mainFrame
-        }
-
-        return NSScreen.screens.first?.visibleFrame ?? NSRect(x: 80, y: 80, width: 1440, height: 900)
-    }
-
-    private static func primaryVisibleFrame() -> NSRect? {
-        if let originScreen = NSScreen.screens.first(where: {
-            abs($0.frame.minX) < 1 && abs($0.frame.minY) < 1
-        }) {
-            return originScreen.visibleFrame
-        }
-
-        return NSScreen.screens
-            .sorted {
-                if $0.frame.minY == $1.frame.minY {
-                    return $0.frame.minX < $1.frame.minX
-                }
-                return $0.frame.minY < $1.frame.minY
-            }
-            .first?
-            .visibleFrame
     }
 }
